First pull of the project.
All the files were reviewed and checked. The review consists mainly
in the comprehension of the code, that is to say, correct functions,
parameters, and explain some actions. The main file added has, for
the moment, two example cases, proximately a new example case going
to be added. The two cases that exist refer to how to (i) run multiple
times with different parameters of the algorithm and (ii) how to get
existing models and fit them.